+++ game/ato/package.py
@@ -187,12 +187,14 @@ class Package(RadioFrequencyContainer):
             FlightType.TARCAP,
             FlightType.BARCAP,
             FlightType.AEWC,
+            FlightType.ISR,
             FlightType.FERRY,
             FlightType.RECOVERY,
             FlightType.REFUELING,
             FlightType.SWEEP,
             FlightType.SEAD_ESCORT,
             FlightType.ESCORT,
+            FlightType.JAMMING,
         ]
         for task in tasks_by_priority:
             if flight_counts[task]: